+++ packages/crdt_lf_devtools_extension/lib/src/application/document_changes/document_changes_cubit.dart
@@ -41,7 +41,8 @@ class DocumentChangesCubit extends Cubit<DocumentChangesState> {
 
   void _setupEventSubscription() {
     _eventStreamSubscription = args.service.onExtensionEvent.listen((event) {
-      if (event.isDocumentChangedEvent) {
+      if (event.isDocumentChangedEvent &&
+          event.documentId == args.document.id) {
         _loadDocumentChanges();
       }
     });